Make boot resolve base assets by default
