+++ packages/runtime/src/index.ts
@@ -24,7 +24,8 @@ export type {
   VsockExecResult,
 } from "./exec.ts";
 export type { ChunkLogEvent, LogEvent, OnLog, PhaseLogEvent } from "./log.ts";
-export { provision, resolveBaseDtb, resolveBaseKernel, resolveBaseRootfs } from "./provision.ts";
+export { resolveBaseDtb, resolveBaseKernel, resolveBaseRootfs } from "./base-assets.ts";
+export { provision } from "./provision.ts";
 export type { ProvisionOptions, ProvisionResult } from "./provision.ts";
 export {
   ensureRootfsImage,
